fix:according_to_comments

+++ database/migrations/2021_01_26_201401_create_appointments_table.php
@@ -21,7 +21,7 @@ class CreateAppointmentsTable extends Migration
     $table->unsignedBigInteger('doctor_id');
     $table->foreign('doctor_id')->references('id')->on('doctors')->onDelete('cascade')
     ->onUpdate('cascade');
-    $table->integer('time');
+    $table->bigInteger('time');
     $table->tinyInteger('duration');
     $table->tinyInteger('status')->default(1);
     $table->text('cancel_reason')->nullable();
